esp runtime without interrupt and exception handling at least for reaching/booting our (finally nearly) empty main function
finalising runtime assembler code: zeroing .bss and registers, allocating
.stack, setting frame pointer then calling _start_rust rust function

_start_rust calls __post_init (which is empty), _setup_interrupts (whic is
not implemented yet), then finally calls hal_main in esp-hal
hal_main calls the #[entry] decorated function, this case our main funtions

runtime needs more segments, sections and labels, so esp32c3.x, memory.x and
linkall.x linker scripts were extended and rwdata.x and stack.x was created

in the runtime assembler there is no _vector_table yet, hence exception trap
handling (_start_trap_rust_hal) and setup of preemptive/prioritised interrupts
(_setup_interrupts) is not implemented yet

to call _setup_interrupts we need an empty function in esp-hal/src/interrupt/
/riscv, but since I skipped the relavant trap handling code in runtime asm
_start_trap_rust_hal did not have to be implemented

the file riscv32.size contains a "size table" with the sizes of all functions
currently the loaded code (.text + .rodata) is 252 bytes

it was a painful experience, and it took me almost a whole day to find out
why this programme could not run,
even though the main function at first was just the empty loop, but this code
should have loaded and run normally, but it didn't

even the debugger (gdb) was not able to run, it always got wrong memory segment
mappings from openocd, which showed that there was some problem with the loaded
image
it turned out that espflash flashed the image incorrectly if there was only one
loadable segment (.text), but if there was a .rodata segment in the code then
espflash mapped the memory segments successfully after loading and the code ran
correctly
so I needed a minimal .rodata segment, but if it was not used, the compiler
optimized it,
to avoid this, I called the usb_uart_tx_one_char ROM function to consume the
bytes from a static byte string, but this also only worked if the string size
was larger than 59 bytes, at which point it was no longer optimized.

then I put this 'static' behind a reference and passed it to another function,
from there I called usb_uart_tx_one_char, this also avoided optimization and
it was enough to use only one static byte (which put 4 bytes in .rodata) to
make the program run

there is a commented version of rom_usb_print in the code, which could be a
simple (unformatted) console print function in the future, which does not
require a usb printer driver to be written

if you don't touch the "mtvec" register, i.e. don't overwrite it with our own
_vector_table, the ROM provides a default "Guru Meditation Error" panic and
exception handler,
so I think this program would be one of the smallest executable codes on this
chip, which of course requires this shrinked runtime