esp-backtrace
I pretty much followed the original implementation of esp-backtrace. Although
I have left out parts that are unnecessary for me, related to the colors,
custom-halt, custom-pre-backtrace and halt-cores features. Also, there is no
difference between println vs defmt features, as it is in esp_println for me
is resolved. Since the panic_immediate_abort option should not be enabled here,
since panic won't even get to the rust_begin_unwind feature, I checked the
difference between defmt and no defmt (core). In this single line panic
generating unwrap test program the difference was not significant, since almost
no readonly (static) string is put into the .rodata segment, instead the many
empty println in the output of the esp-backtrace generates many _defmt_acquire
and _defmt_release pairs. But as the code and the .rodata in it grows the size
difference increases significantly even without panic_immediate_abort.

I had to repair the esp32c3.x linker script because it was linking the .trap
segment incorrectly, and was running into error it exception handlingn in
esp-hal. I dlso eleted the derive(defmt::Format) from the TrapFrame struct and
manually implmented the Format trait, because its output was pretty ugly.

I've used the esp-backtrace crate before, and then I noticed that for defmt,
if the panic was generated in the main function, it didn't show at whichpoint
in main. So I changed the #[inline(always)] decoration of the defmt panic call
 to #[inline(never)]. This way it worked. However, I now realize (I hope) that
the problem is not here, but in esp-backtrace.

After some investigation it appeared that if the backtrace and the
backtrace_internal function called from it are compiled inlined then exactly in
the case of suppress=2 the third address to be shown, i.e. in the case of
suppress=2 exactly the first address to be shown, disappears from the 'result'
array returned by backtrace_internal. The compiler somehow generates different
code, and this value is somehow lost in the optimization. It was hard to figure
this out, because when I traced the contents of this array with println, the
error never occurred. But when I took out the println's, I always got bad
arrays with a missing address. So I had a hard time figuring out that inline
and inline(never) affect the behavior of the code here (which of course was
switched on and off by the explicit println).
I didn't want to get more involved in this, so I put that #[inline(never)] before
backtrace function then everything worked smoothly.

However, this way there was no further need to customize the defmt crate. So I
eliminated the defmt submodule and rewrote its local references from everywhere