Finishing the minimum esp32c3 application
Some finishing touches. Create a main with test cases. Their outputs and their
impact on the code size explained in the file es32c3.size.

This milestone is worth a tag. I will use it as a starting point for my further
development.

I also wrote the first version of the repository readme.