Continuing to implement a minimal embassy application.
The way of working was that I first solved all the errors coming from the
building of the two proc macros written by the previous commit.

But in the structures, only their fields are defined, not the methods where
they could be omitted. Where methods were essential and their return values
were not critical, I left them empty.

This method provides a simple but clear picture of the basic structures and
their relationship to each other.

Since the integrated-timers feature is not defined yet, this also reduces the
complexity of the first view. In embassy-executor, I have chosen the "nightly"
feature. This feature allocates all tasks in statics. To do this, the
impl_trait_in_assoc_type feature in main has to be enabled.

We don't need the interrupt driven executor yet, so I just implemented the
thread executor. I wrote two document files. In the first one, I tried to
analyze expanded macros, in the second one, where and why it is necessary to
define the send and sync traits.