RunQueue(enqueue, dequeue), polls, waker, pender
Implementing the missing code that the two proc macros needed. The top level
run/event loop is in the executor of esp-hal-embassy. This is where "__pender"
defined which sets SIGNAL_WORK_THREAD_MODE AtomicBool to true. The interrupt
that causes the __pender exits from the wfi statement of the loop. Then the
interrupt handler of the task calls the waker passed to the task when it was
polled. The waker calls the __pender function of the Executor associated with
the task. After the interrupt, we find ourselves after the wfi statement. In
the loop, we immediately notice the signal and call (polls) raw::Executor. That
calls (polls) raw::SyncExecutor. The latter traverses all the enqueued tasks
sitting in the RunQueue and calls (polls) the tasks' future functions. That is,
the "main" and "run" functions decorated by the two proc macros.

This code is finally running. As a minimal test, few traces have been added at
"rtos traces" points to the code.

As a minimal test, we have added some traces to the code, in places where there
are "rtos traces". This way we can see that the main task is actually created,
run, and then ececutor is get into wait/idle/wfi state.