Eliminating esp-backtrace. I've reviewed it carefully again, even the extensa part as well.
I've forked esp-hal and modify esp-backtrace with a patch which eliminate to use println feature.

Then I set patches in config.toml for esp-backtrace and all other crates which were coverged back
to their crates.io version before this commit (esp-println, esp-sync, esp-metadata-generated and
esp-config)i to my forked version of esp-hal.

The defmt feature has limited use here, as it gives the location (PanicInfo) where defmt::panic
calls core::panic!(). So defmt::panic does not give the location info of the original panic.
Therefore, it does not need the core formatting snippets.
I previously posted an issue on https://github.com/knurling-rs/defmt. But now I realized that it
was wrong. The problem is not with the captured addresses -- thay are good -- but this missing
original PanicInfo.