While learning/developping esp-rtos, I tried to bring my esp-hal closer to 1.0.0
The origial motivation was to fully understand the interrupt handling in esp-hal
for better understanding the nested interrupts for esp-rtos. So I gradually built
my esp-hal/interrupt module until I reached the 1.0.0 functionality. After that
I should made a building consistency to other modules (clock, gpio, systimer,
timg, etc). Then I finally replaced my last dependency, the esp-ricv-rt module
to its factory version (my forked version of esp-hal at 1.0.0 stage and used as
a patch pointig my locally mirrored repository). The goal was to build this more
and more updated version of esp-hal not only with the gradually build esp-rtos
but also with my original application (ecu-via-elm327).

But as I realised it is impossible, since the version of TrapFrame in the new
version of esp-rsiv-rt (comming from risv-rt) is different from the version used
in my custom esp-wifi. In esp-wifi/preempt_builtin module, the old scheduler for
esp-wifi and the predecessor of esp-rtos uses a full TrapFrame, saving all possible
registers when entering and returning from interrupt. But the new verison of
TrapFrame backups only part of those registers. So the old internal scheduler
is not be able to run with the updated/factory version of esp-risv-rt which
needed by the more and more updated version of esp-hal.

Anyway I will follow this path: learning/developping esp-rtos, learning/updating
the used parts of esp-hal. So, if everything goes well, when I switch to esp-rtos
I can use the esp-hal(.me) from that time, before swithcing to the factory version
more precisely version 1.0.0, with the latter.